+++ src/utils/oauth.ts
@@ -1,3 +1,6 @@
+// src/utils/oauth.ts
+
+// dependency modules
 import { google } from 'googleapis';
 
 const oauth2Client = new google.auth.OAuth2(
