+++ src/utils/prisma.ts
@@ -1,3 +1,8 @@
+// src/utils/prisma.ts
+
+// dependency modules
 import { PrismaClient } from '@prisma/client'
+
 const prisma = new PrismaClient()
+
 export default prisma
